+++ server/src/controllers/index.ts
@@ -1,5 +1,7 @@
 import config from './config';
+import worker from './worker';
 
 export default {
   config,
+  worker,
 };